kafka-producer/requirements.txt
psycopg2-binary==2.9.9
kafka-python==2.0.2